Add check on CNAP if call state changed
BZ: 204181

Currently if the CNAP information is received and sent
after the framework has been notified of an incoming
call, the UI is not updated accordingly.
To check if two calls are the same, only the ID is
compared. This is why a check on the CNAP has also
been added.

Moreover, CacheInfoEntry is not updated when CNAP
changes so an update is forced.

Change-Id: I0bb1e9104c9818ad4c7b9dacc007f9bb3104011f
Category: aosp improvement
Domain: Telephony-Telephony-FWK
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Morgane BUTSCHER <morganex.butscher@intel.com>