Update name display when number is unknown
BZ: 204030

When CNAP information is received and sent
after the framework has been notified of an incoming
call, the UI is not updated accordingly if the number
is unknown.

A member is added to the ContactCacheEntry class
to be able to know when the number is unknown and
therefore when it is allowed to override the stored
name with the CNAP information.

Change-Id: If8df0507c93f40cd95f9983cd3b10bc5541a47f7
Category: aosp improvement
Domain: Telephony-Telephony-FWK
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Morgane BUTSCHER <morganex.butscher@intel.com>